backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
